Query Binding Exploitation high severity Published 8 hours ago in laravel/framework Package  illuminate/database (composer) Affected versions < 6.20.12 >= 7.0.0, < 7.30.3 >= 8.0.0, < 8.22.1 Patched versions 6.20.12 7.30.3 8.22.1 Package  laravel/framework (composer) Affected versions < 6.20.12 >= 7.0.0, < 7.30.3 >= 8.0.0, < 8.22.1 Patched versions 6.20.12 7.30.3 8.22.1 Description Description Laravel versions <6.20.12, <7.30.3 & <8.22.1 contain a query binding exploitation.
If a request is crafted where a field that is normally a non-array value is an array, and that input is not validated or cast to its expected type before being passed to the query builder, an unexpected number of query bindings can be added to the query. In some situations, this will simply lead to no results being returned by the query builder; however, it is possible certain queries could be affected in a way that causes the query to return unexpected results.