Dependabot cannot update ssri to a non-vulnerable version The latest possible version that can be installed is 6.0.1 because of the following conflicting dependencies:
laravel-mix@5.0.9 requires ssri@^7.0.0 via a transitive dependency on cacache@13.0.1
laravel-mix@5.0.9 requires ssri@^6.0.1 via a transitive dependency on cacache@12.0.4
The earliest fixed version is 8.0.1.

Dependabot cannot update is-svg to a non-vulnerable version
The latest possible version that can be installed is 3.0.0 because of the following conflicting dependency:

laravel-mix@5.0.9 requires is-svg@^3.0.0 via a transitive dependency on postcss-svgo@4.0.2
The earliest fixed version is 4.2.2.